Replace hard-coded credit data with dynamic JSON loading
This commit introduces a new JSON file, `credits.json`, to store the data for the credits section of the website. The previous hardcoded HTML in `credit.html` was replaced with JavaScript that dynamically loads the information from the JSON file. This change increases maintainability and flexibility - it's now easier to add or modify credit entries by simply updating the JSON file.