"Updated dark mode colour scheme in multiple CSS and HTML files
Optimized the design of dark mode across pages in the navbar.css, research.html, project.html, credit.html, and index.html files. This was motivated by the need to enhance the overall readability and user experience of the site in dark mode. Specific changes include a darker background for the `body.dark` class, text color adjustments for better visibility and contrast, and specific updates to the .dark-mode-icon class for a consistent and aesthetically pleasing user interface."