Improve image brightness in dark mode
Added a brightness filter to the images in dark mode to ensure they do not appear too bright and strain the user's eyes. Applied the filter specifically on the project-card class images when body is set to dark theme. This enhances the visibility and overall user experience in dark mode.