Reduce gap and add border to badge images
The CSS style for "badge-container" and "github-badge-container" has been updated. Gap between badges has been reduced and a border was added to the images for better visual separation. Additionally, the organization of badges was changed, where CSS3 badge is now before JavaScript, and badge-container margin bottom was reduced from 5px to 2px for better content fit.