Update main logo and make it dark mode compatible
The main logo was updated from "has_logo.svg" to "regional-2023-logo_black.svg" across all HTML files. Additionally, "dark-mode.js" was updated so that the new logo changes to the appropriate color when dark mode is toggled. The logo files for both light and dark modes were also added. The change was necessary to reflect the new branding for 2023 regional events.