Add hover effect to GitHub badge
In order to enhance user experience and to make the badge look more interactive, I've added a hover effect to the GitHub badge in the credit.html page. The changes include a slight drop in the opacity, an application of a shadow for a 3d effect as well as a cursor change to signify a clickable item. Additionally, a bright shadow has been provided for a better visibility on dark backgrounds.