Modify 'credit.html' CSS for dark mode image brightness
Refined the CSS selector in 'credit.html'. Previously, all images in the dark mode had their brightness adjusted to 80%, which was causing inconsistencies in the styling. Now, only images in the '.credit-section' have their brightness adjusted when in dark mode. It improves the website's aesthetics and maintains a consistent design in dark mode.